+++ src/components/tab/main/CommonBlogMain.vue
@@ -1,0 +1,13 @@
+<template>
+<div>CommonBlogMain</div>
+</template>
+
+<script lang="ts">
+export default {
+  name: "CommonBlogMain"
+}
+</script>
+
+<style scoped>
+
+</style>
