+++ src/components/tab/main/commonblogmainadaptor/LiandiMain.vue
