+++ src/components/tab/main/metaweblogmainadaptor/WordpressMain.vue
@@ -3,7 +3,7 @@
 </template>
 
 <script lang="ts" setup>
-import {API_TYPE_CONSTANTS} from "../../../lib/constants/apiTypeConstants";
+import {API_TYPE_CONSTANTS} from "../../../../lib/constants/apiTypeConstants";
 import {ref} from "vue";
 
 const apiType = ref(API_TYPE_CONSTANTS.API_TYPE_WORDPRESS)
@@ -11,7 +11,7 @@ const apiType = ref(API_TYPE_CONSTANTS.API_TYPE_WORDPRESS)
 </script>
 
 <script lang="ts">
-import MetaweblogMain from "./MetaweblogMain.vue";
+import MetaweblogMain from "../MetaweblogMain.vue";
 
 export default {
   name: "WordpressMain",
