+++ src/components/tab/setting/CommonBlogSetting.vue
@@ -1,0 +1,13 @@
+<template>
+<div>CommonBlogSetting</div>
+</template>
+
+<script lang="ts">
+export default {
+  name: "CommonBlogSetting"
+}
+</script>
+
+<style scoped>
+
+</style>
