+++ src/components/tab/setting/metaweblogsettingadaptor/WordpressSetting.vue
@@ -3,9 +3,9 @@
 </template>
 
 <script lang="ts" setup>
-import {API_TYPE_CONSTANTS} from "../../../lib/constants/apiTypeConstants";
+import {API_TYPE_CONSTANTS} from "../../../../lib/constants/apiTypeConstants";
 import {ref} from "vue";
-import {WordpressCfg} from "../../../lib/platform/wordpress/wordpressCfg";
+import {WordpressCfg} from "../../../../lib/platform/metaweblog/config/wordpressCfg";
 
 const apiType = ref(API_TYPE_CONSTANTS.API_TYPE_WORDPRESS)
 const cfg = ref(new WordpressCfg())
@@ -13,7 +13,7 @@ const cfg = ref(new WordpressCfg())
 </script>
 
 <script lang="ts">
-import MetaweblogSetting from "./MetaweblogSetting.vue";
+import MetaweblogSetting from "../MetaweblogSetting.vue";
 
 export default {
   name: "WordpressSetting",
