+++ src/lib/platform/metaweblog/adaptor/wordpressApiAdaptor.ts
@@ -1,0 +1,9 @@
+import {MetaWeblogApiAdaptor} from "../metaWeblogApiAdaptor";
+import {IApi} from "../../../api";
+import {API_TYPE_CONSTANTS} from "../../../constants/apiTypeConstants";
+
+export class WordpressApiAdaptor  extends MetaWeblogApiAdaptor implements IApi {
+    constructor() {
+        super(API_TYPE_CONSTANTS.API_TYPE_WORDPRESS);
+    }
+}
