+++ src/lib/platform/metaweblog/config/wordpressCfg.ts
@@ -1,6 +1,6 @@
-import {MetaweblogCfg} from "../metaweblog/MetaweblogCfg";
-import {POSTID_KEY_CONSTANTS} from "../../constants/postidKeyConstants";
-import {PageType} from "../metaweblog/IMetaweblogCfg";
+import {MetaweblogCfg} from "../MetaweblogCfg";
+import {POSTID_KEY_CONSTANTS} from "../../../constants/postidKeyConstants";
+import {PageType} from "../IMetaweblogCfg";
 
 /**
  * 博客园配置类
