--- src/lib/platform/wordpress/wordpressApiAdaptor.ts
@@ -1,9 +1,0 @@
-import {MetaWeblogApiAdaptor} from "../metaweblog/metaWeblogApiAdaptor";
-import {IApi} from "../../api";
-import {API_TYPE_CONSTANTS} from "../../constants/apiTypeConstants";
-
-export class WordpressApiAdaptor  extends MetaWeblogApiAdaptor implements IApi {
-    constructor() {
-        super(API_TYPE_CONSTANTS.API_TYPE_WORDPRESS);
-    }
-}
