+++ tsconfig.json
@@ -14,7 +14,7 @@
     "skipLibCheck": true,
     "allowJs": true,
     "checkJs": false,
-    "types": ["node", "stripe-v3", "vite/client"]
+    "types": ["node", "vite/client"]
   },
   "include": ["src/**/*.ts", "src/**/*.d.ts", "src/**/*.tsx", "src/**/*.vue", "test/**/*.ts"],
   "references": [{ "path": "./tsconfig.node.json" }],
